Implement pop quiz system after each page. (ISSUE #168)
Adds JavaScript and SCSS to convert quiz data in markdown to an interactive pop quiz that can be played by the user.
Added quiz system. See README for more details or the binary page for an example.
This commit also fixes navigation and JavaScript/SCSS files on Netlify by making all links relative.